+++ src/Sulu/Component/Content/Mapper/ContentMapper.php
@@ -465,7 +465,7 @@ class ContentMapper implements ContentMapperInterface
                     $propertyTranslator->getName('published')
                 );
             }
-        } else {
+        } elseif (isset($state)) {
             $this->changeState(
                 $node,
                 $state,
@@ -1613,13 +1613,18 @@ class ContentMapper implements ContentMapperInterface
     /**
      * {@inheritDoc}
      */
-    public function copyLanguage($uuid, $userId, $webspaceKey, $srcLanguageCode, $destLanguageCodes)
-    {
+    public function copyLanguage(
+        $uuid,
+        $userId,
+        $webspaceKey,
+        $srcLanguageCode,
+        $destLanguageCodes,
+        $structureType = Structure::TYPE_PAGE
+    ) {
         if (!is_array($destLanguageCodes)) {
             $destLanguageCodes = array($destLanguageCodes);
         }
 
-        /** @var Page $structure */
         $structure = $this->load($uuid, $webspaceKey, $srcLanguageCode);
 
         $data = $structure->toArray(true);
@@ -1635,7 +1640,9 @@ class ContentMapper implements ContentMapperInterface
                 $uuid,
                 null,
                 Structure::STATE_TEST,
-                $structure->getIsShadow()
+                $structure->getIsShadow(),
+                $structure->getShadowBaseLanguage(),
+                $structureType
             );
         }